feat: 조회 로직 구현 및 Join fetch 적용으로 최적화 / API 구현

+++ backend/src/main/java/com/deliveranything/domain/delivery/dto/response/TodayDeliveringResponseDto.java
@@ -9,7 +9,7 @@ public record TodayDeliveringResponseDto(
     LocalDateTime now,
     RiderToggleStatus currentStatus,
     Long todayDeliveryCount,
-    Integer todayEarningAmount,
+    Long todayEarningAmount,
     Double avgDeliveryTime
 ) {
